Refactor and add features
Change folder structure to have an app directory.

The root directory will now contain the files that are independent of
the app and soon directories for migrations and unit tests.

Work on models for the database and start using queries to populate
views. Add functionality to encrypt and verify user data when
authenticating.

Created a file to initialize a sqlite database for testing.

Uploading now requires authentication.